Rework to have Vector3 on his own files
